requirements.txt
@'
aiogram>=3.22.0,<4.0.0
aiohttp==3.12.15
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
'@ | Set-Content -Encoding UTF8 requirements.txt
